refactor: Move UI components from src/features/ui to src/components/ui
🔄 COMPONENT REORGANIZATION:

✅ MOVED:
- src/features/ui/*.tsx → src/components/ui/*.tsx
- All 9 common UI components relocated to more logical location

✅ UPDATED:
- All import paths: @/features/ui/ → @/components/ui/
- CLAUDE.md documentation to reflect new structure
- Directory structure examples in documentation

✅ PRESERVED:
- src/features/ directory with .gitkeep for future feature components
- Git history for moved files (using git mv)

🎯 RATIONALE:
- features/ should be for business logic/feature-specific components
- components/ui/ is more appropriate for pure UI components
- Clearer separation between features and reusable UI elements
- Aligns with conventional React project structures

✅ VERIFICATION:
- Build passes successfully ✓
- All imports resolved correctly ✓
- Documentation updated ✓

This creates a more intuitive and maintainable component organization.
