+++ src/app/loading.tsx
@@ -1,7 +1,7 @@
 // reference
 // url:https://zenn.dev/catnose99/articles/19a05103ab9ec7
 
-import { LoadingSpinner } from '@/features/ui/loading-spinner'
+import { LoadingSpinner } from '@/components/ui/loading-spinner'
 
 export default function Loading() {
   return (
