+++ src/components/ui/category-icon.tsx
