+++ src/components/ui/image-wrapper.tsx
